Fix a garbage collection bug
If we use Rcpp::Nullable, then it equivalent of assigning SEXP, i.e., copying a pointer, and thus the R function is not reference counted. If the passed in R function is a temporary variable, then it got released when returning to R from the C++ constructor, and cause a problem when the garbage collector is run.