Make Agent::leave return the agent
When the Agent::leave method is called in an R event handler, the agent  is passed by reference. So, when it is removed from the population, the smart pointer holding the agent has no user, and so the agent is freed. This is no problem if we are using it for the death of an agent, but it poses a problem if we want to move the agent from one population to another in the event handler. With this function returning the smart pointer, the agent will not be freed.