Changed the handling of Agent ID and index in the population
Agent should have a unique ID in a simulation. The current implementation uses the population index as agent ID. But now agents can move between population, so their ID changes with it. Now the ID is unique and assigned when an agent is attached to the population. The index is managed but the parent population.

The change does affects the public c++ interface and the R interface.